move test class
